refactor: move toAddress finding function to utils file
